requirements.txt
boto3
pytest
requests
PyYAML
mock
pytest-cov
python-dotenv